Move Character Types into a dedicated package
